Feature/rent bounded context

+++ src/main/java/com/techonwheels/wheelmanager/rent/application/commands/CreateVehicleCommand.java
@@ -1,10 +1,12 @@
 package com.techonwheels.wheelmanager.rent.application.commands;
 
-import com.techonwheels.wheelmanager.rent.domain.valueobjects.Brand;
-import com.techonwheels.wheelmanager.rent.domain.valueobjects.Status;
+import com.techonwheels.wheelmanager.rent.infrastructure.valueobjects.Brand;
+import com.techonwheels.wheelmanager.rent.infrastructure.valueobjects.Status;
 import lombok.Builder;
 import lombok.Data;
 import org.axonframework.modelling.command.TargetAggregateIdentifier;
+
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -14,7 +16,7 @@ public class CreateVehicleCommand {
     private String vehicleName;
     private String imageURL;
     private String description;
-    private Double price;
+    private BigDecimal price;
     private Double qualification;
     private Status status;
     private Brand brand;
