+++ src/main/java/com/techonwheels/wheelmanager/rent/application/events/VehicleCreatedEvent.java
@@ -1,11 +1,13 @@
 package com.techonwheels.wheelmanager.rent.application.events;
 
-import com.techonwheels.wheelmanager.rent.domain.valueobjects.Brand;
-import com.techonwheels.wheelmanager.rent.domain.valueobjects.Status;
+import com.techonwheels.wheelmanager.rent.infrastructure.valueobjects.Brand;
+import com.techonwheels.wheelmanager.rent.infrastructure.valueobjects.Status;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -16,7 +18,7 @@ public class VehicleCreatedEvent {
     private String vehicleName;
     private String imageURL;
     private String description;
-    private Double price;
+    private BigDecimal price;
     private Double qualification;
     private Status status;
     private Brand brand;
