+++ src/main/java/com/techonwheels/wheelmanager/rent/domain/aggregates/VehicleAggregate.java
@@ -2,14 +2,16 @@ package com.techonwheels.wheelmanager.rent.domain.aggregates;
 
 import com.techonwheels.wheelmanager.rent.application.commands.CreateVehicleCommand;
 import com.techonwheels.wheelmanager.rent.application.events.VehicleCreatedEvent;
-import com.techonwheels.wheelmanager.rent.domain.valueobjects.Brand;
-import com.techonwheels.wheelmanager.rent.domain.valueobjects.Status;
+import com.techonwheels.wheelmanager.rent.infrastructure.valueobjects.Brand;
+import com.techonwheels.wheelmanager.rent.infrastructure.valueobjects.Status;
 import org.axonframework.commandhandling.CommandHandler;
 import org.axonframework.eventsourcing.EventSourcingHandler;
 import org.axonframework.modelling.command.AggregateIdentifier;
 import org.axonframework.modelling.command.AggregateLifecycle;
 import org.axonframework.spring.stereotype.Aggregate;
 import org.springframework.beans.BeanUtils;
+
+import java.math.BigDecimal;
 
 @Aggregate
 public class VehicleAggregate {
@@ -18,7 +20,7 @@ public class VehicleAggregate {
     private String vehicleName;
     private String imageURL;
     private String description;
-    private Double price;
+    private BigDecimal price;
     private Double qualification;
     private Status status;
     private Brand brand;
