+++ src/main/java/com/techonwheels/wheelmanager/rent/domain/dto/response/VehicleResponse.java
@@ -1,11 +1,13 @@
 package com.techonwheels.wheelmanager.rent.domain.dto.response;
 
-import com.techonwheels.wheelmanager.profile.domain.data.Customer;
-import com.techonwheels.wheelmanager.rent.domain.data.VehicleType;
-import com.techonwheels.wheelmanager.rent.domain.valueobjects.Brand;
-import com.techonwheels.wheelmanager.rent.domain.valueobjects.Status;
+import com.techonwheels.wheelmanager.profile.infrastructure.data.Customer;
+import com.techonwheels.wheelmanager.rent.infrastructure.data.VehicleType;
+import com.techonwheels.wheelmanager.rent.infrastructure.valueobjects.Brand;
+import com.techonwheels.wheelmanager.rent.infrastructure.valueobjects.Status;
 import lombok.Builder;
 import lombok.Data;
+
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -14,7 +16,7 @@ public class VehicleResponse {
     private String vehicleName;
     private String imageURL;
     private String description;
-    private Double price;
+    private BigDecimal price;
     private Double qualification;
     private Status status;
     private Brand brand;
